Lockhammer: Document average lock depth metric
Add an explantion in the comments to describe the purpose of the auxiliary
metric "average lock depth".  This metric is an algorithm-specific method
for reporting an estimate of contention level.  For example, a queue-based
algoirthm might report the queue depth at the time of each acquisition
indicating the number of lock acquisitions which must occur before the newly-
enqueued thread can acquire the lock.

Change-Id: Id9ae677bf2f68c8803a656ba3ac2ff873d0ffe16
Signed-off-by: Lucas Crowthers <lucasc.qdt@qualcommdatacenter.com>