Lockhammer: Add per-thread realtime metric
In addition to the global acquire time (realtime) and the per-thread
scheduled time metrics add an additional metric which measures the
average real time between successful acquires on a particular thread.
Please be aware that this commit changes the output format in order
to include the new metric and may necessitate changing and output
parsing scripts.

Fixes #12

Change-Id: I9978172938b13887b46285cb9a10cd1d73afc68b
Signed-off-by: Lucas Crowthers <lucasc.qdt@qualcommdatacenter.com>