lockhammer: Create a calibrated wait loop
This change enables a calibrated wait loop that can measure in the
ns range on non-turbo enabled systems.  Turbo can throw off the
calibration phase if we're not careful.

Additionally this patch adds additional command line parsing to
differentiate between asking for time based waits for the critical
section and parallel sections and the old # of iterations method.

Lastly this patch re-implements both the x86 and Arm wait loops as
inline assembly.

Change-Id: I9d9c6258aa30804407eb33ff95a79727a2f26a4f
Signed-off-by: Geoffrey Blake (Geoffrey.Blake@arm.com)