lockhammer -o and -i may have feature interaction when -i != 1
If all values in -o are zero, then lockhammer will fallback to -i
interleave pinning scheme. However, if -o are not all zero and -i
is not 1, then there are cases where two threads may pin to the
same core because -o and -i scheme conflicts.

We will not fix this in either lockhammer.c or test_lockhammer.py
because we can use -o and -i at the same time to minute tuning
pinning order, e.g. use -i as the base and use zero in -o as the
mask to mask out the core we don't care. And specify the sequence
as a positive integer we want change.

Sample: 4-core machine [core 0, core 1, core 2, core 3]:

-i 2 should yield the following thread pinning sequence:
[0, 2, 1, 3]

-i 2 -o 0,0,3,1 has this sequence, whereas 0 serving as mask:
[0, 2, 3, 1]

As a result, we have to change default lh_test_cfg.yaml -i to 1
because -i will not conflict with -o 0,1,2,3 anymore.