lockhammer: add option to run benchmark for a specified number of timer ticks
While lockhammer synchronizes the start of the locking threads, it does not
synchronize the completion of the threads.  Each thread acquires and releaes
the lock for a specified number of iterations in a loop.  However, some threads
may finish all of their iterations earlier than others, which leads to less
contention for the threads who have not finished.  This results in a
performance measurement that does not describe a full all-thread contention.

This patch provides an option to use a different measurment loop that
terminates after a specified number of timer clock ticks has passed.  To
provide a way to avoid a situation where reading the timer clock tick has high
overhead, the measurement loop only samples the timer after a flag-specified
number of lock/release iterations.

-O ticks	number of timer clock ticks (CNTVCT_EL0 or TSC)
-I inner_iters	inner iterations of measurement loop in between timer reads