keep only one .gitignore file
In the original commit, the .gitignore files are empty files for making
the directory structure.  However, those directories are now populated
by other files, and the Makefile will make the build directories, so the
empty .gitignore files are no longer needed.

Add a top-level .gitignore that covers the entire source tree.

Change-Id: I0c5bd3f4f43a7978141f94c682e1c2e01e5f23b0