run-tests.sh
This script runs a set of lockhammer executables from an embedded list
of test/build/variants against another a list of critical/parallel
durations across a number of detected CPUs to do a core-scaling study.

The result of each test invocation is stored to a json file.

The lists in the script are to be edited by the user.

The script is meant to be easily modified for different scaling studies.

Change-Id: I6da249bd015b46e139dac51d338a4c790d507785