Always measure TSC frequency on x86
The TSC frequency is not easy to obtain.  The methods that this patch
replace do not always work.  However, the TSC frequency is used for
blackhole and test durations, so an accurate number is important.

This patch improves the TSC frequency measurement routine to support
sub-second samples.  On every run of the program, the frequency is
measured for 3 samples of 0.1 seconds, dropping the highest and lowest.

Change-Id: Ifc62210b8da586fbaebf0a3f22d72bc2eebda92b
Signed-off-by: jty2 <46021128+jty2@users.noreply.github.com>