measure.c - pick median of 5 evaluate_blackhole searches
The blackhole delay loop effects the critical and parallel times.  It
takes an argument that affects the duration of the delay.  Generally,
the larger the number, the longer the delay.  However, the delay is not
strictly monotonically increasing with the value of the argument.

To improve reproducibility, when a time is specified in nanoseconds, do
the blackhole calibration 5 times and select the median result.

These individual calibrations are visible when --more-verbose is used.

Change-Id: Id94d6139f4c12430b5e854983e941d99b0d46418
Signed-off-by: James Yang <james.yang@arm.com>