tbb_spin_rw_mutex: pure reader CPU list
The tbb_spin_rw_mutex test-specific -m cpumask flag for specifying pure
readers is limited to affecting 64 software threads because it is a
bitmask.  Moreover, it is inconvenient to match the corresponding bit
that represents a thread to the logical CPU number.

This patch replaces -m cpumask with -o cpulist, where the CPUs are
specified by numbers joined by commas or a colon, just like the main
Lockhammer pinorder -o cpulist flag.  If a specified CPU is in the
pinorder cpulist of an experiment, then it will be a pure reader.

For example, this invocation will run with 8 threads, with cpus 3, 5,
and 7 as pure readers.

    lh_tbb_spin_rw_mutex ... -t 8 -- -o 3,5,7

Change-Id: Iaa3c8d3c0adf0a03561151036449fa0f3ceea8f2
Signed-off-by: James Yang <james.yang@arm.com>