ext/linux/include/lk_atomics.h: add "cc" to clobber list for cmpxchg
For x86_64, cmpxchgl sets CC flags, so add it to the atomic_cmpxchg_*
clobber list.  This should not have an effect if "cc" is implicit for
every asm().

xchgw doesn't set CC flags, but the kernel has "cc" in the clobber
list because it uses the same macro as xadd, which does set flags.
So, don't remove "cc" for authenticity reasons.

Change-Id: I2c20c9f713cdcb17f16522aada4dffd0eb29fa7c
Signed-off-by: James Yang <james.yang@arm.com>