Makefile: templatize recipes, add asm and dump targets, use info
Reduce the verbosity of the the specific lock recipes
by using a template.

Add targets to generate assembly files (.s).

Add targets to generate disassembly using objdump.
All of the locks for a given variant can be generated using
    make alldump
All of the locks for allvariants can be generated using
    make alldumpvariants
alldumps and alldumpsvariants are aliases for these targets.

Also, instead of using echo to display progress, use $(info ).

Also, misc cleanup.

Change-Id: I2964b3948b271c55430910225260f4453bb3ea00
Signed-off-by: James Yang <james.yang@arm.com>