Move interleaving into -t and add -o ranged cpulist
For -o pinorder, sequential CPU list ranges m-n[:skip] can be
specified by joining the starting and ending CPUs with a '-', e.g.:

  -o 1,2,4-7        # use CPU 1,2,4,5,6,7

The listed pinorder string is shown in the output report.

The :skip parameter selects only every skip'th CPU in the range.

  -o 8-15:2	    # use CPU 8,10,12,14

As with individual CPUs, CPU ranges can be combined using commas:

  -o 0-15:2,16-23   # CPU 0,2,4,6,8,10,12,14,16,17,18,19,20,21,22,23

The interleaving flag (-i / --interleave-pinning) is removed and
the functionality is subsumed into -t num_threads[:interleave].
interleave=0 is the default (meaning to choose the next available
logical CPU in the CPU affinity mask).  interleave values greater
than 0 algorithmically calculate the logical CPUs to use.

Otherwise, -t num_threads has the same functionality as before because
--interleave-pinning 0 was the default.

Each pinorder string given by -o or num_threads string given by -t
is stored to the json, except the latter is also prefixed with "t"
to distinguish it from a pinorder of a single CPU number.  The
cpuorder_filename given by -C is also stored to the json.

Change-Id: Ia53412984f666fed39587b2e4cb6997c53f9baad
Signed-off-by: James Yang <james.yang@arm.com>