lockhammer: Add -T tag to store a tag string to results JSON
This patch adds "-T tag", which stores the given tag as a string to the
results JSON.  A tag is useful for discerning different builds of the
benchmark program from the same source code, or different hardware or
operating system configurations not otherwise tracked by lockhammer.

Note: Previously, -T was used as the short flag for --run-limit-seconds,
which has the workload run for the given duration in sconds.  A previous
patch renamed the -T flag to -D

Change-Id: I242bae1892e04d16a45b0a1c4df0d1b80c1c5d32
Signed-off-by: James Yang <james.yang@arm.com>